Fix #3617
I don't know what lua files besides core are going to be loaded when stuff in lua_common.cpp is called, so I dropped the lua metatable stuff in core.

Very much open to changing that if there's a more 'proper' place for it.

As far as the code itself goes, instead of double-storing like celmin suggested, I felt that mapping attempts at accessing or modifying indexes 'x' or 'y' to 1 or 2 respectively was cleaner, as there's no risk of one piece of code modifying 'x' and another then trying to use a value from index 1.